requirements.txt
numpy>=1.19.0
pandas>=1.1.0
torch>=1.7.0
biopython>=1.78
pysam>=0.16.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.23.0
tqdm>=4.50.0
scipy>=1.5.0
